+++ lib/lightning_web/live/channel_request_live/show.ex
@@ -55,7 +55,11 @@ defmodule LightningWeb.ChannelRequestLive.Show do
         <LayoutComponents.header current_user={@current_user}>
           <:breadcrumbs>
             <LayoutComponents.breadcrumbs>
-              <LayoutComponents.breadcrumb_project_picker project={@project} />
+              <LayoutComponents.breadcrumb_project_picker
+                project={@project}
+                current_user={@current_user}
+                access_root={@access_root}
+              />
               <LayoutComponents.breadcrumb_items items={[
                 {"History", ~p"/projects/#{@project}/history"},
                 {"Channels", ~p"/projects/#{@project}/history/channels"}