+++ lib/lightning_web/live/workflow_live/edit.ex
@@ -95,7 +95,11 @@ defmodule LightningWeb.WorkflowLive.Edit do
         <LayoutComponents.header current_user={@current_user}>
           <:breadcrumbs>
             <LayoutComponents.breadcrumbs>
-              <LayoutComponents.breadcrumb_project_picker project={@project} />
+              <LayoutComponents.breadcrumb_project_picker
+                project={@project}
+                current_user={@current_user}
+                access_root={@access_root}
+              />
               <LayoutComponents.breadcrumb_items items={[
                 {"Workflows", "/projects/#{@project.id}/w"}
               ]} />